+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ### 🛑 Breaking changes 🛑
 
-- (Splunk) `receiver/discovery`: Emit entity events for discovered endpoints with log_endpoints: true ([#4684](https://github.com/signalfx/splunk-otel-collector/pull/4684))
+- (Splunk) `receiver/discovery`: Update the component to emit entity events 
+  - Emit entity events for discovered endpoints with log_endpoints: true ([#4684](https://github.com/signalfx/splunk-otel-collector/pull/4684))
+  - Ensure active endpoints emitted as entity events periodically ([#4684](https://github.com/signalfx/splunk-otel-collector/pull/4684))
 
 ### 💡 Enhancements 💡